Updates EvalTree and MPI distribution. Still debugging!
EvalTree update (making subtrees map to contiguous segments of the
parent tree) has been debugged and tested so it nominally works
now.  The point of this update was to allow for a cleaner (and
faster) MPI distribution.  Namely, computed values can be inserted
directly into the final result array eliminating lots of copying
(which results in a ratcheting memory demand).  This streamlined
computation has been implemented for bulk_fill_dprobs and the
functions it calls (but no hprobs functions yet).  This meant
adding several functions to mpitools.py and redefining the
wrtFilters parameter to be a dictionary of *slices* instead of
a dictionary of lists of integers.   Cache computation functions
also take *slices* instead of interger lists now, which allows
for increased speed by sacrificing flexibility that wasn't (and
shouldn't) be needed.

Currenlty, there are *many* debug statements and large sections
of commented out debug and old-for-reference code (in evaltree.py,
gscalc.py, core.py, customlm.py, and mpitools.py).