Updates to get tests to pass on mpitest branch, plus minor mods.
Running unit tests revealed several boundary cases which needed
to be addressed, in particular regarding the creation and
splitting of EvalTree objects.  Secondly, the aggressive custom-LM
algorithm would fail for some cases (paramterized gates in ptic)
and so the custom LM algorithm has been tuned back to be less
aggressive (higher damping) and needs further tweaking to realize
full performance potential.

At this point all tests pass (before merge w/develop), and the only
updated reference file is the gateset estimate for the model
selection test, which I expect is different b/c of the low accuracy
in the test and updated LM algorithm.  In any case, the model
selection test is more a proof that the code runs than a test of
expected results at this point.

Also at this point, the MPI distribution is evolved to the point
that we expect close to the desired linear scaling until
computations are dominated by things like the gauge optimization
and computing the number of gauge params (so decent overall
multi-processor scaling).