Add assert statement in rb_results.plot() function
This assert statement checks whether the .plot() function
call has a correct argument ("prim_or_cliff") prior to starting to generate
the plot. If the argument is not correct, an AssertionError
is raised.

Previously, the figure was created and then a series
of if/then statements were executed. As a result, if
the argument was incorrect, part of the figure axes
would be drawn.

This commit tidies that up some.