Change plot colors, markers, and lines in rb_results.plot()
This commit introduces some new colors, markers,
and line styles to the rb_results.plot() function.

Using the "Set1" colormap, I chose a blue-ish color
for the markers, and an orange-ish color for the lines.
In this way, the two should be pretty distinct (complementary
colors), and a bit easier to see. I also changed the linewidth
to be a bit thicker, and made the markers a bit larger.

Finally, I re-ordered the plotting calls to plot the line
first, and then the markers. That way the latter sit on top
of the former, which looks a little bit nicer.