Set figures to have larger default size
I don't know exactly what matplotlib uses
as the default figure size when you initialize
a Figure object. It was a bit too small, so
I declared a figsize of (8, 4), meaning it's
twice as wide as it is tall. This size appears
to work well with the other formatting choices
I made.