Subtracts 1 from the lengths used in RB fits & plots.
By subtracting 1 from the sequence lengths before performing the
RB fits and plots we do *not* include final inverting Clifford
(as desired).  For primitive/other gate label sets, we still just
subtract 1 for now, since computing the actual number of gates
that correspond to the final Clifford would be some work and
shouldn't  matter to anyone.