Adds collisionAction argument & functionality to data set objects.
In light of proper RB protocol needing to keep data from duplicate
gate sequences separate, a new argument 'collisionAction' has been
added to the constructor of DataSet (and MultiDataSet) objects.
By default it is set to "aggregate" which causes the add-data
functions of DataSet to add the supplied data to an existing gate
sequence, which has been the only behavior thus far.  Now, by
setting 'collisionAction' to "keepseparate" the DataSet's add-data
methods will append a "#<number>" gate label to the end of
existing gate sequences (starting with <number> == "1" and going up)
and thereby store the data separately.  To access the data, one
may query the gate string + "#1", for example, or use the new
'get_row' and 'set_row' methods which take the optional integer
argument "occurance" which defaults to 0.  Machinery to track
and preserve the collisionAction of DataSets within a
MultiDataSet has also been added, though it may not be needed
much since MultiDataSets store *static* DataSets.  Thus, it
only comes into play when you make a non-static copy of one
of the data sets stored within a MultiDataSet.

This update also adds 'collisionAction' arguments to io's
load_dataset and load_multidataset, as well as to
generate_fake_data.

Furthermore, the RB functionality has been updated to make
proper use of collisionAction, so that do_randomized_benchmarking
expects a DataSet with collisionAction == "keepseparate" (or with
no duplicated clifford sequences).  The tutorial notebook has
been updated to generate a fake data set accordingly.