Adds "ftol" convergence check to custom_leastsq.
By adding a convergence condition based on the relative sum-of-
squares being below a certain threshold, the custom LM routine
now will exit earlier in certain slow-converging problems thereby
reducing its run time.  This behavior is similar to SciPy's
leastsq routine, and now in almost all cases the custom routine
finishes within at most twice the time required by the SciPy routine,
and with values not as much lower than the SciPy values.  So overall
this makes the LM routine more similar to the SciPy one in terms of
convergence time and accuracy.

The case of parametrized gates (in testDrivers.py) is still
particularly difficult to converge, and deservers further
investigation in the future.