Fixes a bug in do_long_sequence due to updated argument list.
When io.load_dataset was given a new collisionAction argument,
the call to load_dataset in do_long_sequence_gst was not updated
and thereby passed a verbosity printer object as the collition
action.  This was only discovered by trying to pickle a resulting
DataSet, which failed due to Comms not being picklable.  DataSet
has been updated to check for a valid collisionAction.