Removes fickle low-memory tests and several bug fixes.
This commit comments out parts of MC2GST and MLGST tests
(in testCore.py) that were meant to invoke memory control (i.e.
tree-splitting) by giving a very small memory limit.  However,
the current memory estimation uses as a base point the current memory
in use by the process, which can vary somewhat from system to
system and instance to instance.  This makes it nearly impossible to
set a fixed value of the limit that always induces memory control
without sometimes making the effective memory limit negative, which
now causes an exception to be raised (and the test to fail).

Also updated are:
- EigenvalueParameterizedGate's construction when a matrix has
  multiple unit-eigenvalue eigenvectors has been made more
  robust (test was failing on OS X).
- Fiducial selection algorithm used ar[[indices]] syntax,
  where indices is a tuple of indices, and this was failing on
  OS X, perhaps due to some python-implementation detail.  In any
  case, it's more clear to write array[list(indices)], which I
  believe is the intent of the initial version, and so this has
  been updated.