Fixes a isinstance integer comparison to work with numpy ints too.
Discovered by Marcus, there existed a bug in the GateString class:
When multiplying a GateString by a number, it is asserted that the
multplier be a non-negative integer BUT it only checked for
Python ints and not numpy ints.  This is problematic especially
when constructing gate string lists, which catch assertion errors
silently as a signal to ignore a particular string.  Now this
has been fixed and all integer types work. [Perhaps make an
explicit test for this!]