Updates do_long_sequence_gst to compute starting gateset serially.
In some contexts it was observed that gauge optimizing an LGST
gateset can vary (slightly) among different processors (in an
MPI context).  Thus, to ensure that the iterative GST algorithm
is seeded with *exacty* the same starting gate set, the
do_long_sequence_gst driver function now computes the starting
point (often LGST) on just the rank 0 processor and then broadcasts
it to the others (instead of each processor independently computing
the starting point).