Updates reference reports so DEEP TESTS PASS on Erik's laptop.
Current report tex files were compared to the reference reports,
and have been updated so that now (by fiat) the "deep" latex-
comparting tests pass.  Here's an overview of what did and did
not need to be updated in the reference reports:

- non-TP chi2 reports (reports A & B) were nearly identical
  to the reference reports "on file".  Only 3rd- or 4th- digit
  differences, which are just numerical error (the last ref.
  reports were generated using a different machine, and even
  on Erik's current machine there are differences of this order
  from run to run (see below).  The one exception is that some
  of the error bars in full_reportBv3.tex were erroneously large
  in the *existing* reference report and these have been fixed
  now.

- the non-markovian error bar test report (report E) had similar
  minor differences to the reference as for reports A & B.

- TP logL reports (reports C & D) had substantial updates because
  the *existing* reference reports didn't optimize the logL as
  well as the current ones do (likely due to the new LM algorithm
  out-performing scipy's leastsq in many circumstances).  This
  causes many of the estimated values to change, but remain the
  same order of magnitude.  The previous reference-report error
  bars in these cases could be erroneously large, and this has
  been fixed in the new reports - sometimes an update with
  *different* orders of magnitude.

When running the report tests multiple times in a row (on Erik's
OSX laptop) the latex files would have some variation in the last
(4th) digit, so reportBaseCase.py has been modified to reduce the
precision used in the test report generation.  This ameliorates
the issue, but doesn't make it dissappear, as there's still a
seemingly random sign that the eigensolver gives to eigenvectors
and values close to zero in scientific notation are not consistent
across repeated runs.  This may be due to OSX libraries not being
as "stable" as the linux counterparts?  The logL or chi2 values
do *not* change, suggessting that the numerical differences may
be arising at the gauge optimization (or later processing, e.g.
gate decomposition) stage.  For now, we'll just accept the fact
that the reference reports may differ slightly from the ones
generated - and we'll still call that OK.