Adjusts general reports "snap" scaling to avoid clipping.
A recent change to the general report was to keep individual color
scales for the error generator box plots but to "snap" these scales
to an existing scale that was within an order of magnitude of the
scale induced by the data.  This could cause clipping if the scale
was snapped to one with a smaller range, so now we only snap to
scales with *larger* ranges.  Still a work in progress...