Fixes bugs in intial new-gauge-opt implementation & updates tests.
The new gauge optimization functions ('gaugeopt_*') have been
tested and the tests have been updated accordingly so that at
least most of the tests pass now.  Logic for setting a GateSet's
'default_gauge_group' member has been added, but docstrings are
still missing.

Something to note is that if, in the case when a GateSet is
optimized using all frobenius metrics, the normalization is
changed this causes slightly different resulting GateSets --
so the "special case" when both gatesMetric and spamMetric
are "frobenius" has been added to gaugeopt_to_target.  This
special case seems logical in that the objective function is
then a single call to GateSet.frobeniusdist.