Fixes minor bug in C-J isomorphism & adds related updates.
There was a mix-up in by-col vs. by-row vectorization when
computing the Choi matrix (pyGSTi assumes vectorization is by-row
and the isomorphism code assumed by-column).  This has been fixed
now.  Ramifications were minor, in that previous Choi matrices had
some elements that were incorrect by a minus sign, but in a way
that does not affect the eigenvalues, which are the aspect of
Choi matrix of primary importance for analysis.

Adds a new 'fast_jamiolkowski_iso_std' function which, as its
name implies, computes the Choi matrix in the standard (mx unit)
basis using a fast algorithm.  This new function is now utilized
by the 'sums_of_negative_choi_evals' function to perform its
evaluation more quickly.  This function has also been updated to
correctly infer the gate basis from the passed-in GateSet (that
it didn't before was a bug affecting gauge optimization to CP
in cases when the gates were *not* in the Gell-Mann basis,
e.g. 2Q-GST).

The 'sum_of_negative_choi_evals' has an added 'itemWeights'
argument so that the sums of different gates' eigenvalues can
be weighted differently (potentially useful in gauge optimization).

Debug statements for testing have been left in but commented out,
and will be removed soon.  Unit tests have been updated to check
that fast_jamiolkowski_iso_std gives the same results as
jamiolkowski_iso.