Fixes and updates new general report tables.
Minor enhancements and bug fixes to the recently added figures
and tables to the appendix of the general report (dealing with
comparisons to per-gate and CPTP gauge optimized gatesets).
Explanatory text has been added to Table captions, and CPTP gauge
optimization is limited to 100 iterations so it doesn't take
forever.