Add complex support to vector-matrix conversion
Both  the gmvec_to_stdmx and stdvec_to_stdmx tools in tools.basistools cast the input vector as a float. This is fine for common basis choices, but can be problematic for others (e.g., qutrit Gell-Mann). This update adds a keyword argument to optionally not cast the vector as a float, which avoids this problem. The default behavior is unchanged.

Signed-off-by: John Gamble <jkgambl@sandia.gov>