Alters LindbladParameterizedGate parameterization improving performance.
Previously, LindbladParameterizedGate parameterized the positive-
definite "non-hamiltonian" Lindblad coefficient matrix by its
unique Cholesky decomposition matrix which has a positive diagonal.
The diagonal elements were thus parameterized by the square of
certain "gate parameters".  This added and unnecessary complexity
proved to burden the numeric optimizations during tests, and it
was found preferable to allow the diagonal elements to be negative
as well (thus losing the uniqueness of the Cholesky decomp but
retaining the critical positive definiteness of the coeff matrix).
Now the diagonal elements are in one-to-one correspondence with
gate parameters.