Updates constructor of LindbladParameterizedGate.
The __init__ method of LindbladParameterizedGate now takes
"gateMatrix" and "unitaryPrefator" arguments which can be used
to initialize a gate from an arbitrary CPTP gate matrix.  Before
this, __init__ just stored the given matrix as what is now the
unitary_prefactor member, which, as the name implies, should have
been a unitary map.