Cleans up TravisCI debugging after (seemingly) fixing the issue.
Ok, a little background:  the TravisCI builds suddenly started
failing when trying to "pip install cvxopt" because "umfpack.h"
wasn't found (probably a new cvxopt dependency?).  These and other
header files are contained in the "libsuitesparse-dev" apt package,
so "apt-get libsuitesparse-dev" was added to the install script, and
after figuring out where this actually put "umfpack.h" a few flags
were added to the "pip install cvxopt" so that it could find umfpack.h.
This got the build a little farther, but it then failed when trying to
link with "-lsuitesparseconfig", whose corresponding library apparently
doesn't exist anywhere.  At the end of the day, I tried copying a
dummy .so file (lapack actually) to "libsuitesparseconfig.so" and that
cause the linking to succeed (indicating that the build didn't actually
need anything from "-lsuitesparseconfig").  So thinks seem to be working
again now - that is, the tests are running again.