Adds more arguments to LindbladParameterizedGate constructor.
New arguments "cptp", "onlydiag" and "truncate" have been added
to LindbladParameterizedGate.__init__.  The first two are just
stubs for later implementation, but "truncate" now allows the
user to specify whether negative eigenvalues of the other-params
matrix (when projecting onto Lindblad terms) get truncated to zero
or cause an error to be raised.

Also, some updates to the constructor have been made to get rid
of complex casting to real warnings.