Catches LinAlgError from matrix_rank calls in GateSet.
When computing the non-gauge projector checks are performed to
test the matrix rank.  In some cases the SVD doesn't converge
resulting in a LinAlgError (I don't understand why yet, but this
occured during a test using LindbladParameterizedGates).  Now,
when this happens a warning is issued but the exception is caught
and the function returns.