Updates germ selection, making it more flexible and verbose.
The "forceSingletons" argument of germ selection routines has been
replaced with a more general "force" argument, which can be set to
the string "singletons" to mimic the previous forceSingletons=True
case.  "force" can also be set to a list of GateStrings, enabling
the user to demand that *any* arbitrary list of germs be present in
the final set.

Also, better stdout printing has been added to `build_up_breadth`,
for more verbose progress reporting (particuarly for 2Q germ
selection).