Adds support for "cptp" and "onlydiag" in LindbladParameterizedGate.
This commit wires up the "cptp" and "onlydiag" arguments passed
to the LindbladParameterizedGate constructor.  These arguments
switch whether CPTP-ness is enforced and whether all or only the
diagonal Lindblad terms are used to parameterize the gate.
Correctness of the derivative has been checked with respect to a
finite difference derivative.