Adds cptp_penalty_factor support to core GST and driver routines.
The long-lost "cptp_penalty_factor" argument, which years ago
controlled a CPTP penalty term in do_mlgst, do_mc2gst, and iterative
GST routines, has been revived.  Previously a "not-implemented"
error would be raised if cptp_penalty_factor was not equal to zero.
Now this functionality as been re-implemented as a vector of the
trace-norms of each gate's choi matrix, which when they are
constrained to TP is identical to the negativity of the gate (it's
non-CPTP-ness).  The form of this penalty has the nice property
that it is differetiable, so all jacobians remain analytic.  A
small note is that currently the penalty term does not include
anything to require the SPAM portions of the gateset to be physical
(eigenvalues of effect vectors being in [0,1] and prep vectors
 corresponding to positive density matrices).

The do_long_sequence_gst driver has been plumbed to allow
advanced options "cptpPenaltyFactor" and "check" to control this
L1-like CPTP penalty term (and jacobian checking, which was just
for convenience during debugging) from a high level.

There's currently still a bunch of debugging statements and test
code commented out in core.py, which should be removed soon, and
as a side effect gatetools.py has a new "tracenorm" function.
Computation of the penalty term and its jacobian have not been
optimized for memory or speed yet -- this is just an initial
version.