Adds memory and performance optimizations to cptp_penalty_factor mode.
After profiling and tuning on a cluster, do_mlgst and do_mc2gst have
been modified so that the jacobian computation for the L1 CPTP
penalty factor is performed more efficiently in both time and
memory.  It now requres negligible overhead to time and memory
required for the jacobian computation when an L1 CPTP penalty
is enabled.

Also updates testCore.py to test CPTP penalty for do_mc2gst
and to not expect a NotImplementedError (since it's implemented
now!).