Removes debug safeguard that clipped other-Mx evals to >= 5e-4.
During debugging, the coefficient matrix of "other" Lindblad terms
for a given gate was clipped so it's eigenvalues were at least
5e-4 (not zero) to ensure if was *very* positive definite (this was
during debugging why were we getting non-CP gates from a pos-def
other-mx, which turned out to be a normalization issue).  I forgot
to later set this 5e-4 back to 0, where it should have been - so
this is done in this commit.  This *only* affected the "LND CPTP"
line in Table 17 of the general report appendix, which previously
would show a *worse* fit even when staring with CPTP gates (so
where it should show the *same* fit).  This error should remedied.