Adds support for unitary gates and gauge optimization.
Changes the "onlydiag" argument of LindbladParameterizedGate's
constructor to the more general "nonHamTerms" which can take
values "all" (same as onlydiag=False before), "diag" (same as
onlydiag=True before), and "none".  When set to "none" only the
Hamiltonian-type generators are included, resulting in a
parameterization over all unitary gates.  This "unitary gate"
functionality is used by the new "UnitaryGaugeGroup" class which
allows one to gauge optimize only over a unitary gauge group. In
particular, LindbladParameterizedGate.transform now allows a
unitary gate transformation by a unitary gauge group element.

Note that the performance of LindbladParameterizedGate.transform
is not good, as the implementation is meant to be the most basic,
fast, and fool-proof way to get it right.