Adds initial implementation of "linear-response" error bars.
Following Robin's notes, this commit adds an additional type of
Hessian projection (set via a Results object's "hessianProjection"
parameter): "linear respose".  When reports are generated from a
Results object having this projection type, error bars are computed
for only a selected subset of tables by computing the response of
the log-likelihood to a forcing term. (This method avoids computation
of the full Hessian and is thus suitable to 2Q-GST and other larger-
Hilbert-space GST.)  The ConfidenceRegion object has been modified
to hold the arguments for and run the new '_do_mlgst_base' core.py
routine which adds the necessary forcing-term functionality to
the public-API routine 'do_mlgst'.