Adds option to specify None as filename to create_general_report_pdf.
One can now specify None as the filename argument of a Results
object's create_general_report_pdf in order to compute and cache
many of the values needed to make the report but not actually
output any files.  This is intended for use with clusters, whereby
the report quanties (e.g. linear reponse error bars) may be
computed on the cluster using many CPUs and the reported generated
(quickly) on a local machine.