Fixes bug in L1 CPTP penalty computation.
I forgot the implicit squaring of the least squares algorithm so
the CPTP pentalty  was actually an L2 penalty not an L1 (doh!).
This has been fixed now, so that that the 'cptp_penalty_factor' in
do_mc2gst, do_mlgst, and related routines multiplies an *L1* term
equal to the sum of the tracenorms of the gates' choi matrices.