Adds option (and new default) for "type" of error generator in reports.
The error_generator(...) function in gatetools.py now can be passed
a "typ" argument which can presently be set equal to either "logTiG"
or "logG-logT" to specify how the error generator should be computed.

This option has been propagated up to the Results class, whose options
(i.e. the resultsObj.options member) now include "errgen_type".  Setting
this value, which defaults to "logG-logT" (a *new* method of error
generator computation) affects the way error generators in reports
are computed as well as the descriptive text, namely the latex
formula - so the text in the reports remains consistent with the
calculations for any value of resultObj.options.errgen_type.