Adds a check that all procs get the same forcefn_grad term.
A broadcast is performed to makes sure all processors receive the
same forcefn_grad argument.  When they don't a warning is printed
but no error raised - the root proc's value is used regardless.
This has shown itself to be an issue when computing diamond norm
error bars and CVXOPT fails on only a subset of the processors -
so if no sync was performed different procs would receive *very*
different forcefn_grad arguments to _do_mlgst_base(...).