Changes the method of computing the "sign" matrix in cptp jacobian.
In cptp-penaltized MLGST and MC2GST the sign of the choi matrix is
computed to form a jacobian.  The way this was done previously
was numerically inaccurate leading to inaccurate jacobians and
assertion failures (for hermiticity of relevant matrices).  Using
the svd to compute the sign matrix fixes this issue.