Adds "depolarize()" as a Gate and SPAMVec method.
Gate and SPAMVec (-derived) objects are now responsible for knowing
how to depolarize themselves, via a new "depolarize" method.  This
allows differently parameterized gates/vectors to be depolarized
via GateSet.depolarize, which calls the contained Gate and SPAMVec
depolarize functions *when a uniform depolarization is requested*.
(If using a random depolarization that can be different for different
basis elements, then, as before, GateSet.depolairze constructs a
FullyParameterizedGate.) In particular, this allows the use of
the 'depolarizeStart' do_long_sequence_gst advanced option to
function properly (i.e. *not* change the parameterization).