Fixes a bug in gather_slices and adds gather_slices_by_owner.
A bug introduced at the Hessian-row-distribution upgrade (a use
of 'iaxis' where 'axis' was needed) has been fixed and now
everything seems to be working fine as tested in a multi-cpu
enviroment.  The new gather_slices_by_owner isn't currenlty used
for anything, but was during testing different Hessian distribution
ideas and it seems general-purpose enough to keep around.