Replaces std2Q_XYCNOT effect strings with new "aggregatable" strings.
'std2Q_XYCNOT.effectStrs' now refers to the new effect fiducial strings,
which have been tested (along with the recently added length-65 germ
set) to give the desired 1/L scaling for 2Q-GST.  The old effect
strings are still accessible via 'std2Q_XYCNOT.legacy_effectStrs'.

To keep naming more uniform, the old set of 71 germs, whcih was
accessible via the '.germs71' member, is now '.legacy_germs'.

The motivation for these "new" germ and fiducial sets is:
- all sets (prep fids, meas fids, & germs) are able to be used
  for 1Q-GST on either of the qubits, which we're calling
  "aggregate-able".
- the fiducial sets are short.
- the germ set contains all the single gates.

We find that the germ and fiducial "scores" are nearly identical
to those of the old sets, and we've seen that the new sets give
rise to the desired 1/L scaling when using several randomly
depolarized (but not too much) & rotated gatesets to generate
fake (Markovian) data.