Fixes bug whereby CPTP contraction did not contract prep vecs.
When calling contract(gs, "CPTP",...) when useDirectCP == True,
the preparation SPAM vectors would not be contracted to TP (ie
their first element set to 1/gateDim**0.25).  If useDirectCP were
False or if just contracting to "TP", the prep vecs would be
contracted correctly -- this was a bug for this particular case.