Adds warning instead of error for complex generator projections.
When the result of a projection is complex (when it's expected to
be real) a warning is printed instead of a ValueError being thrown
and the computation proceeds with the real-part of the projection.
This nicer behavior seems warranted as we tweak the projection
calculations within reports.

This commit also updates some import statements to conform to
the pattern used throughout pyGSTi.