Adds checks for None in generate_fiducials to gracefully exit.
When no valid fiducial set is found, optimize_integer_fiducials_slack
returns None.  Previously, when this occured, an exception would
result when trying to compute the score of the fiducial set (which was
None, though a list of gate strings was expected).  Now the score is
only computed when a non-None result is obtained, fixing the issue.