Improves logG-logT generator.  Asserts that error-gens should be real.
With the addition of the 'real_matrix_log' function for finding
the real logarithm (when possible) of a matrix, the logG-logT
generator is now real whenever the G and T matrices have evenly-
degenerate negative eigenvalues.  While this may not always be
possible, it seems to hold in most cases of interest (i.e. target
gates up to 2Q) so far, so it is now considered an error when an
error generator is not real (this addresses the issue of having to
plot a complex error generator).  Now, 'error_generator' raises
a ValueError when the computed error generator is not real.