Adds meta-data final section to full and general reports.
A new meta-data section containing tables of relevant pyGSTi
parameters & options, as well as current system information (e.g.
versions of various relevant software) has been added at the end
of the full (1-qubit) report and the general (1+ qubits) report.
The motivation for this addition is capture in each report more
details regarding how GST was actually run, and what the software
environment is.  There will be more information added to this
section as time goes on; currently we just dump all the parameters
and options of the Results object (which should be expanded in the
future).