Adds "comment" member to DataSet and MultiDataSet & carries to reports.
DataSet and MultiDataSet can now be passed a comment string in their
constructors.  This comment is saved & loaded along with the data,
and allows users to attach some specific information to the data.
When a DataSet (or MultiDataSet) is read in from a text file, the
single-# (commented) lines of the preamble (before the data) are
taken to be the comment.  In reports, the "dataset overview" table
now includes the DataSet comment (if it exists), by adding a new
row for each line of the comment.  Overall, this update allows
users to add information to their data set text files that will
be propagated through to the reports.