Adds "maxWeight" argument to LindbladParameterizedGate.
The new "maxWeight" argument to the constructor of
LindbladParameterizedGate allows the user to limit the basis used
in the Lindbladian representation to just those elements with a
weight (i.e. number of non-identity 1Q factors) that is less than
a given value.  This currently is only allowed for Gates in the
Pauli-product basis, as this is the only basis at present which is
neatly factorize-able into 1Q pieces.  This update also generalizes
the LindbladParameterizedGate implementation so that is no longer
assumes (and requires the gat to be in) the Pauli product basis --
except when using the aforementioned 'maxWeight' functionalify.

The update is primarily passing down "maxWeight" through error-
generator functions in gatetools.py to a call to pp_matrices (in
basistools.py) which now accepts "maxWeight" as an argument.