Adds commented-out code showing how to enable MLGST on each iter.
Within do_iterative_mlgst a commented-out block is added that (if
un-commented) would perform MLGST after MC2GST on *each* iteartion.
This was used for the testing of accuracy scaling and may in the
future be added as an optional mode for do_iterative_mlgst.  From
our accuracy testing, it was observed that just doing MLGST instead
of MC2GST at each iteration was unreliable (as had been seen in
previously too).