Fixes bug in general report: "relative eigenvalues" were incorrect.
In the "relative eigenvalue" table of the general report (the one
containing the polar plots of the relative eigenvalues), the first
text column of the table, which was supposed to be listing the
relative eigenalues numerically, was actually listing the eigenvalues
of the generator (which by default is the logarithm of the "relative
gate").  This has been fixed now.