Propagates gateLabelAliases up to top level; fixes bug in direct plots.
The ability to specify a gateLabelAliases dictionary (which defines
a set of replacement rules when accessing only the dataset) has
been added to many functions and propagated all the way up to
do_long_sequence_gst where it can be specified as the advanced
option 'gateLabelAliases'.  Corresponding updates to all report
generation functionality has also been done, to that running
create_full_report_pdf with all appendices succeeds when a
'gateLabelAliases' dictionary is supplied.

Also, this commit fixes a bug in the direct-X plotting routines
which has gone uncaught for some time (functions would run fine
but produce a blank plot).  This was due to incorrectly applying
gatestring, fiducial pair, and dataset filtering when computing
the probability matrix for the ('GsigmaLbl',) string used as a
placeholder in direct-X computations.  The plotting tutorial
notebook is updated so it now contains this plot correctly.