Updates LindbladParameterizedGate to make it more flexible & faster.
The LindbladParameterizedGate constructor now takes separate
"hamiltonian" and "non-hamiltonian" basis arguments, which, in
addition to being one of the basis names (e.g. "pp" or "gm") can
be a custom basis (list of matricies) to use.  The "maxWeight" and
"nonHamTerms" arguments have been removed, and a
"nonham_diagonal_only" argument has been added, which specifies
whether to use all or just the diagonal non-hamiltonian terms.

These top-level changes have been propagated to the gatetools.py
functions lindblad_error_generators and lindblad_errgen_projections,
which do a lot of the heavy lifting.

This is still a work in progress: it's very lightly tested and
none of the doc strings are up to date yet (but are marked as
needing updating).