Updates nongauge projector construction; adds nullspace_qr function.
The construction of the nongauge projector (in order to compute the
number of gauge or non-gauge parameters) was problematic in 3Q-GST
tests because 1) it used a lot of memory and 2) the nullspace(...)
routine, which uses the SVD to compute the nullspace, would fail
with a LinAlgError "SVD could not converge" (and also used lots of
memory).

The get_nongauge_projector GateSet function has been modified to
remove a call to numpy.concatenate (reducing the function's memory
usage) and to use the new nullspace_qr(...) routine which uses a
QR decomposition to compute the nullspace.  Supposedly, the QR
method is less accurate, but this seems fine, given that we're usually
just interested in the integer number of gauge parameters, and
it doesn't fail in the 3Q-GST case.  So, until/unless we find
that the added accuracy of the SVD nullspace routine is needed,
the get_nongauge_projector routine will soley use nullspace_qr.