Fixes assertion that final_view creates a numpy view.
Previously, the test for "is B a view of A" was

assert(B.base is A),

but this seems to not always be the case, and in some circumstances
(maybe when A itself is a slice?) the correct test is:

assert(B.base is A.base)

instead. Now we assert that either condition holds, i.e.

assert(B.base is A or B.base is A.base)