Adds support for setting gaugeOptParams == False to disable G.O.
do_long_sequence_gst could (and can) be passed the argument
'gaugeOptParams=False' to disable gauge optimization.  However,
this didn't play nicely with downstream report generation which
assumed at various point a dictionary of gauge optimization
arguments/parameters (or a list of such dictionaries).  Not report
generation is updated to handle the case when gauge optimization is
disabled (gaugeOptParams == False).