Fixes CNOT and CPHASE gates in std2Q gatesets, adds intial XYI versions.
CNOT and CPHASE gates were errant because of a 1Q phase error in the
implementation -- these gates are *not* simply CX(pi) and CZ(pi)
rotations as implemented in pyGSTi!  This commit fixes the standard
gatesets by explicitly construting the correct gates, but the germs
still need to be reviewed to see if they're still AC.

Also, std2Q_XYICNOT and std2Q_XYICPHASE std gatesets have been added,
which include the identity "Gii" gate.  Currently the germs for these
sets are just the germs of the non-Gii sets plus the single "Gii"
germ (thinking that it will amplify all of its own errors) but this
needs to be checked (along with the fact that the non-Gii germs need
to be checked themselves).