Updates report "pages" color boxplot to be more robust.
Previously, an error could occur when a "page" of the divided
color box plot (in the general report) contains no non-repeated
(i.e. visible) plaquettes.  This condition essentially leads to
an empty color box plot and a "None" instead of a valid ReportFigure
object.  Now this potential None value is properly handled by
simply not creating the plot (instead of raising an error that
None doesn't have a save_to method).