Adds commented-out tweaks to custom-LM routine.
During the course of debugging some ill-behaved data some
hacks/tweaks were found to be somewhat useful in getting the LM
routine to converge *slightly* more robustly.  These tweaks are
not well motivated from an algorithmic point of view, and are
therefore commented out now, but it seems good to leave them
around for similar debugging in the future.