Adds a memory check for computing num_nongauge_params in GST iters.
Functions do_mlgst and do_mc2gst compute the gateset's number of non-gauge
parameters in order to construct an expected logL or chi2 value and
when printing status messages to the user.  On large Hilbert spaces
finding the number of gauge parameters can be memory and computationally
intensive, so now a check is performed:  if the memory required to
compute the number of non-gauge parameters exceeds 10% of the user-
specified memory limit, the computation is not performed and the
total number of gate set parameters are used as the number of
"model parameters" instead (a message is printed indicating when
this shortcut is used).